requirements.txt
opencv-python
numpy
pytorch
torchvision
random
matplotlib
thop
